refactor: clean-up old code and update docs
