+++ .github/workflows/a2a-docker-build.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - '*'
     tags:
-      - 'v*'
+      - '*'
 
 env:
   IMAGE_NAME: cnoe-io/agent-argocd
