--- .github/workflows/acp-docker-build.yml
@@ -1,60 +1,0 @@
-name: ACP Docker Build and Push
-
-on:
-  push:
-    branches:
-      - '*'
-    tags:
-      - 'v*'
-
-env:
-  IMAGE_NAME: cnoe-io/agent-argocd
-
-jobs:
-  build-and-push:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      packages: write
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Log in to GitHub Container Registry
-        uses: docker/login-action@v3
-        with:
-          registry: ghcr.io
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Extract metadata for Docker
-        id: meta
-        uses: docker/metadata-action@v4
-        with:
-          images: ghcr.io/${{ env.IMAGE_NAME }}
-          tags: |
-            type=raw,value=acp-latest,enable=${{ github.ref == 'refs/heads/main' }}
-            type=ref,event=branch,prefix=acp-
-            type=ref,event=tag,prefix=acp-
-            type=sha,format=short,prefix=acp-
-
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v3
-
-      - name: Build and Push ACP Docker image
-        uses: docker/build-push-action@v6
-        with:
-          context: .
-          file: ./build/Dockerfile.acp
-          push: ${{ github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags/') }}
-          tags: ${{ steps.meta.outputs.tags }}
-          labels: ${{ steps.meta.outputs.labels }}
-          platforms: linux/amd64,linux/arm64
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-          build-args: |
-            AGENT_NAME=argocd
