+++ Makefile
@@ -13,16 +13,15 @@ MCP_SERVER_DIR ?= mcp_$(AGENT_NAME)
 
 ## -------------------------------------------------
 .PHONY: \
-  build setup-venv activate-venv install run run-acp run-client \
-  langgraph-dev help clean clean-pyc clean-venv clean-build-artifacts \
-  install-uv install-wfsm verify-a2a-sdk evals \
-  run-a2a run-acp-client run-a2a-client run-curl-client \
-  run-mcp run-mcp-client test registry-agntcy-directory \
-  build-docker-acp build-docker-acp-tag build-docker-acp-push build-docker-acp-tag-and-push \
-  build-docker-a2a build-docker-a2a-tag build-docker-a2a-push build-docker-a2a-tag-and-push \
-  run-docker-acp run-docker-a2a \
-  check-env lint ruff-fix \
-  help add-copyright-license-headers
+	build setup-venv activate-venv install run \
+	langgraph-dev help clean clean-pyc clean-venv clean-build-artifacts \
+	install-uv verify-a2a-sdk evals \
+	run-a2a run-a2a-client run-curl-client \
+	run-mcp run-mcp-client test registry-agntcy-directory \
+	build-docker-a2a build-docker-a2a-tag build-docker-a2a-push build-docker-a2a-tag-and-push \
+	run-docker-a2a \
+	check-env lint ruff-fix \
+	help add-copyright-license-headers
 
 ## ========== Setup & Clean ==========
 
@@ -72,10 +71,6 @@ install: setup-venv ## Install Python dependencies using Poetry
 install-uv:        ## Install uv package manager
 	@$(venv-run) pip install uv
 
-
-install-wfsm:      ## Install workflow server manager (wfsm)
-	curl -sSL https://raw.githubusercontent.com/agntcy/workflow-srv-mgr/refs/heads/install-sh-tag-cmd-args/install.sh -t v0.3.1 | bash
-
 ## ========== Build & Lint ==========
 
 build: setup-venv  ## Build the package using Poetry
@@ -89,10 +84,6 @@ ruff-fix: setup-venv ## Auto-fix lint issues with ruff
 
 ## ========== Run ==========
 
-run-acp: setup-venv ## Deploy ACP agent via wfsm
-	@$(MAKE) check-env
-	@$(venv-run) wfsm deploy -m ./$(AGENT_PKG_NAME)/protocol_bindings/acp_server/agent.json --envFilePath=./.env --dryRun=false
-
 run-a2a: setup-venv ## Run A2A agent with uvicorn
 	@$(MAKE) check-env
 	@A2A_PORT=$$(grep A2A_PORT .env | cut -d '=' -f2); \
@@ -104,42 +95,24 @@ run-mcp: setup-venv ## Run MCP server in SSE mode
 
 ## ========== Clients ==========
 
-run-acp-client: setup-venv ## Run ACP client script
-	@$(MAKE) check-env
-	@$(venv-run) uv run client/acp_client.py
-
 run-a2a-client: setup-venv ## Run A2A client script
 	@$(MAKE) check-env
-	@$(venv-run) uv run client/a2a_client.py
+	@$(venv-run) uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
 
 run-mcp-client: setup-venv ## Run MCP client script
 	@$(MAKE) check-env
-	@$(venv-run) uv run client/mcp_client.py
+	@$(venv-run) uvx https://github.com/cnoe-io/agent-chat-cli.git mcp
 
-run-curl-client: setup-venv ## Run shell-based CURL client
-	@$(MAKE) check-env
-	@$(venv-run) ./client/client_curl.sh
-
+## ========== LangGraph Development ==========
 langgraph-dev: setup-venv ## Run LangGraph dev mode
 	@$(venv-run) langgraph dev
 
+## ========== Evals ==========
 evals: setup-venv ## Run agentevals with test cases
 	@$(venv-run) uv add agentevals tabulate pytest
 	@$(venv-run) uv run evals/strict_match/test_strict_match.py
 
 ## ========== Docker ==========
-
-build-docker-acp:            ## Build ACP Docker image
-	docker build -t $(AGENT_DIR_NAME):acp-latest --build-arg AGENT_NAME=argocd -f build/Dockerfile.acp .
-
-build-docker-acp-tag:        ## Tag ACP Docker image
-	docker tag $(AGENT_DIR_NAME):acp-latest ghcr.io/cnoe-io/$(AGENT_DIR_NAME):acp-latest
-
-build-docker-acp-push:       ## Push ACP Docker image
-	docker push ghcr.io/cnoe-io/$(AGENT_DIR_NAME):acp-latest
-
-build-docker-acp-tag-and-push: ## Tag and push ACP Docker image
-	@$(MAKE) build-docker-acp build-docker-acp-tag build-docker-acp-push
 
 build-docker-a2a:            ## Build A2A Docker image
 	docker build -t $(AGENT_DIR_NAME):a2a-latest -f build/Dockerfile.a2a .
@@ -152,36 +125,6 @@ build-docker-a2a-push:       ## Push A2A Docker image
 
 build-docker-a2a-tag-and-push: ## Tag and push A2A Docker image
 	@$(MAKE) build-docker-a2a build-docker-a2a-tag build-docker-a2a-push
-
-run-docker-acp: ## Run the ACP agent in Docker
-	@$(MAKE) check-env
-	@AGENT_ID=$$(grep CNOE_AGENT_$$(echo $(AGENT_NAME) | tr a-z A-Z)_ID .env | cut -d '=' -f2); \
-	AGENT_PORT=$$(grep CNOE_AGENT_$$(echo $(AGENT_NAME) | tr a-z A-Z)_PORT .env | cut -d '=' -f2); \
-	ACP_AGENT_IMAGE=$$(grep ACP_AGENT_IMAGE .env | cut -d '=' -f2 || echo ""); \
-	LOCAL_AGENT_PORT=$${AGENT_PORT:-10000}; \
-	LOCAL_AGENT_IMAGE=$${ACP_AGENT_IMAGE:-ghcr.io/cnoe-io/$(AGENT_DIR_NAME):acp-latest}; \
-	echo "========================================================================"; \
-	echo "==                     ACP AGENT DOCKER RUN                           =="; \
-	echo "========================================================================"; \
-	echo "Using Agent Image : $$LOCAL_AGENT_IMAGE"; \
-	echo "Using Agent ID    : $$AGENT_ID"; \
-	echo "Using Agent Port  : localhost:$$LOCAL_AGENT_PORT"; \
-	echo "========================================================================"; \
-	echo "==               Do not use uvicorn port in the logs                  =="; \
-	echo "========================================================================"; \
-	docker run -p $$LOCAL_AGENT_PORT:8000 -it \
-		-v $(PWD)/.env:/opt/agent_src/.env \
-		--env-file .env \
-		-e AGWS_STORAGE_PERSIST=False \
-		-e AGENT_MANIFEST_PATH="manifest.json" \
-		-e AGENTS_REF='{"'$$AGENT_ID'": "$(AGENT_PKG_NAME).graph:graph"}' \
-		-e AGENT_ID=$$AGENT_ID \
-		-e AIOHTTP_CLIENT_MAX_REDIRECTS=10 \
-		-e AIOHTTP_CLIENT_TIMEOUT=60 \
-		-e API_HOST=0.0.0.0 \
-		$$LOCAL_AGENT_IMAGE
-
-# Run Docker container for A2A agent
 
 run-docker-a2a: ## Run the A2A agent in Docker
 	@A2A_PORT=$$(grep A2A_PORT .env | cut -d '=' -f2); \
@@ -208,7 +151,7 @@ test: setup-venv build ## Run tests using pytest and coverage
 ## ========== AGNTCY Directory ==========
 
 registry-agntcy-directory:  ## Push agent.json to AGNTCY registry
-	@dirctl hub push outshift_platform_engineering/$(AGENT_DIR_NAME) ./$(AGENT_PKG_NAME)/protocol_bindings/acp_server/agent.json
+	@dirctl hub push outshift_platform_engineering/$(AGENT_DIR_NAME) ./$(AGENT_PKG_NAME)/protocol_bindings/agntcy_directory/agent.json
 
 ## ========== Licensing & Help ==========
 
@@ -219,3 +162,4 @@ help: ## Show this help message
 	@echo "Available targets:"
 	@grep -E '^[a-zA-Z0-9_-]+:.*?## .*$$' $(MAKEFILE_LIST) | \
 		awk 'BEGIN {FS = ":.*?## "}; {printf "  \033[36m%-30s\033[0m %s\n", $$1, $$2}' | sort
+
