+++ README.md
@@ -9,14 +9,13 @@
 [![Super Linter](https://github.com/cnoe-io/openapi-mcp-codegen/actions/workflows/superlinter.yml/badge.svg)](https://github.com/cnoe-io/openapi-mcp-codegen/actions/workflows/superlinter.yml)
 [![Unit Tests](https://github.com/cnoe-io/agent-argocd/actions/workflows/unit-tests.yml/badge.svg)](https://github.com/cnoe-io/agent-argocd/actions/workflows/unit-tests.yml)
 
-[![ACP Docker Build and Push](https://github.com/cnoe-io/agent-argocd/actions/workflows/acp-docker-build.yml/badge.svg)](https://github.com/cnoe-io/agent-argocd/actions/workflows/acp-docker-build.yml)
 [![A2A Docker Build and Push](https://github.com/cnoe-io/agent-argocd/actions/workflows/a2a-docker-build.yml/badge.svg)](https://github.com/cnoe-io/agent-argocd/actions/workflows/a2a-docker-build.yml)
 ---
 
 ---
 
 - 🤖 **ArgoCD Agent** is an LLM-powered agent built using the [LangGraph ReAct Agent](https://langchain-ai.github.io/langgraph/agents/agents/) workflow and [MCP tools](https://modelcontextprotocol.io/introduction).
-- 🌐 **Protocol Support:** Compatible with [ACP](https://github.com/agntcy/acp-spec) and [A2A](https://github.com/google/A2A) protocols for integration with external user clients.
+- 🌐 **Protocol Support:** Compatible with [A2A](https://github.com/google/A2A) protocol for integration with external user clients.
 - 🛡️ **Secure by Design:** Enforces ArgoCD API token-based RBAC and supports external authentication for strong access control.
 - 🔌 **Integrated Communication:** Uses [langchain-mcp-adapters](https://github.com/langchain-ai/langchain-mcp-adapters) to connect with the ArgoCD MCP server within the LangGraph ReAct Agent workflow.
 - 🏭 **First-Party MCP Server:** The MCP server is generated by our first-party [openapi-mcp-codegen](https://github.com/cnoe-io/openapi-mcp-codegen/tree/main) utility, ensuring version/API compatibility and software supply chain integrity.
@@ -66,22 +65,28 @@ Run the agent in a Docker container using your `.env` file:
 ```bash
 docker run -p 0.0.0.0:8000:8000 -it \
   -v $(pwd)/.env:/app/.env \
-  ghcr.io/cnoe-io/agent-argocd:a2a-latest
+  ghcr.io/cnoe-io/agent-argocd:a2a-stable
 ```
 
 ### 3️⃣ Run the Client
 
-Use the [agent-chat-cli](https://github.com/cnoe-io/agent-chat-cli) to interact with the agent:
+Use the [agent-chat-cli](https://github.com/cnoe-io/agent-chat-cli) to interact with the agent in **A2A mode** using one of the following methods:
 
-```bash
-uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
-```
+- **Using Docker (no local Python environment required):**
+
+  ```bash
+  docker run -it --network=host ghcr.io/cnoe-io/agent-chat-cli:stable
+  ```
+
+- **Using a local Python environment:**
+
+  ```bash
+  uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
+  ```
+
+Choose the method that best fits your setup.
 
 ## Quick Demos
-
-### AGNTCY ACP Demo
-
-![acp_docker_terminal_demo](https://github.com/user-attachments/assets/9f090ce4-87f3-4bc7-8857-2ea4647187d5)
 
 ### Google A2A Demo
 
@@ -92,11 +97,11 @@ uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
 ```mermaid
 flowchart TD
   subgraph Client Layer
-    A[User Client ACP/A2A]
+    A[User Client A2A]
   end
 
   subgraph Agent Transport Layer
-    B[AGNTCY ACP<br/>or<br/>Google A2A]
+    B[Google A2A]
   end
 
   subgraph Agent Graph Layer
@@ -115,6 +120,28 @@ flowchart TD
   D --> E --> F --> E
 ```
 
+## 🔄 Sequence Diagram
+
+```mermaid
+sequenceDiagram
+  participant User as User Client (A2A)
+  participant A2A as Google A2A Transport
+  participant Agent as LangGraph ReAct Agent
+  participant MCP as Langchain MCP Adapter
+  participant MCPServer as ArgoCD MCP Server
+  participant APIServer as ArgoCD API Server
+
+  User->>A2A: Sends request (A2A Protocol)
+  A2A->>Agent: Forwards request
+  Agent->>MCP: Calls tool (MCP Adapter)
+  MCP->>MCPServer: Invokes MCP tool
+  MCPServer->>APIServer: Calls ArgoCD API
+  APIServer-->>MCPServer: Returns API response
+  MCPServer-->>MCP: Returns tool result
+  MCP-->>Agent: Returns tool output
+  Agent-->>A2A: Sends agent response
+  A2A-->>User: Returns response
+```
 ---
 
 ## ✨ Features
@@ -122,7 +149,7 @@ flowchart TD
 - 🤖 **LangGraph + LangChain MCP Adapter** for agent orchestration
 - 🧠 **Azure OpenAI GPT-4o** as the LLM backend
 - 🔗 Connects to ArgoCD via a dedicated [ArgoCD MCP agent](https://github.com/cnoe-io/agent-argocd/tree/main/agent_argocd/protocol_bindings/mcp_server)
-- 🔄 **Multi-protocol support:** Compatible with both **ACP** and **A2A** protocols for flexible integration and multi-agent orchestration
+- 🔄 **Multi-protocol support:** Compatible with **A2A** protocol for flexible integration and multi-agent orchestration
 
 ## Local Development Setup
 
@@ -253,19 +280,21 @@ For more details, see the [official getting started guide](https://argo-cd.readt
 
 ### 3️⃣ Run the A2A Client
 
-To interact with the agent in **A2A mode**:
+To interact with the agent in **A2A mode**, you can use one of the following methods:
 
-```bash
-uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
-```
+- **Using Docker (no local Python environment required):**
 
-or
+  ```bash
+  docker run -it --network=host ghcr.io/cnoe-io/agent-chat-cli:stable
+  ```
 
+- **Using a local Python environment:**
 
-```bash
-make run-a2a-client
-```
+  ```bash
+  uvx https://github.com/cnoe-io/agent-chat-cli.git a2a
+  ```
 
+Choose the method that best fits your setup.
 
 **Example Interaction**
 
@@ -309,7 +338,7 @@ agent_argocd
 │   ├── __init__.py
 │   ├── a2a_server
 │   │   └── agent_executor.py
-│   ├── acp_server
+│   ├── agntcy_directory
 │   │   └── agent.json
 │   └── mcp_server
 │       ├── __init__.py
@@ -404,9 +433,8 @@ See [MAINTAINERS.md](MAINTAINERS.md)
 
 - [LangGraph](https://github.com/langchain-ai/langgraph) and [LangChain](https://github.com/langchain-ai/langchain) for agent orchestration frameworks.
 - [langchain-mcp-adapters](https://github.com/langchain-ai/langchain-mcp-adapters) for MCP integration.
-- [AGNTCY Agent Connect Protocol(ACP)](https://docs.agntcy.org/pages/syntactic_sdk/connect.html)
-- [AGNTCY Agent Gateway Protocol(AGP)](https://docs.agntcy.org/pages/messaging_sdk/agp-howto.html)
-- [AGNTCY Workflow Server Manager (WFSM)](https://github.com/agntcy/workflow-srv-mgr) for deployment and orchestration.
+- [AGNTCY Slim](https://docs.agntcy.org/pages/messaging_sdk/slim-index.html)
 - [Model Context Protocol (MCP)](https://modelcontextprotocol.io/) for the protocol specification.
 - [Google A2A](https://github.com/google/A2A/tree/main)
 - The open source community for ongoing support and contributions.
+
