+++ agent_argocd/__main__.py
@@ -5,7 +5,7 @@ import click
 import httpx
 from dotenv import load_dotenv
 
-from agent_argocd.protocol_bindings.a2a_server.agent import ArgoCDAgent # type: ignore[import-untyped]
+from agent_argocd.agent import ArgoCDAgent # type: ignore[import-untyped]
 from agent_argocd.protocol_bindings.a2a_server.agent_executor import ArgoCDAgentExecutor # type: ignore[import-untyped]
 
 from a2a.server.apps import A2AStarletteApplication
