+++ agent_argocd/protocol_bindings/agntcy_directory/agent.json
