+++ agent_argocd/state.py
@@ -1,4 +1,4 @@
-# Copyright CNOE Contributors (https://cnoe.io)
+# Copyright 2025 CNOE
 # SPDX-License-Identifier: Apache-2.0
 
 from enum import Enum
