+++ pyproject.toml
@@ -15,7 +15,6 @@ dependencies = [
     "a2a-python>=0.0.1",
     "a2a-sdk>=0.2.1",
     "agentevals>=0.0.7",
-    "agntcy-acp>=1.3.2",
     "click>=8.2.0",
     "langchain-anthropic>=0.3.13",
     "langchain-core>=0.3.60",
